+++ CHANGELOG.md
@@ -19,6 +19,7 @@ operations.
 
 ### Changed
 
+- concord-server: return `404` when the secret doesn't exist;
 - concord-server: fix handling of empty inventory query params.
 - concord-server: fixed the repository webhook registration when
 project or repository is created or updated.